cocoa: Don't return unknown event types as `None` from `poll_events()`.
This situation would arise easily in Servo, because we use
`wakeup_event_loop()` a lot, which is internally implemented by posting
`NSApplicationDefined` events to the event loop, which are counted as
unknown events for the purposes of `poll_events()`. This would cause the
events to look like:

   * NSApplicationDefined
   * NSApplicationDefined
   * ...
   * NSApplicationDefined
   * NSMouseMoved
   * NSApplicationDefined
   * ...

And we wouldn't see the `NSMouseMoved` event for a long time. With this
patch, however, in this situation we skip over the
`NSApplicationDefined` events and return the `NSMouseMoved` right away.